API分離作業1
