+++ src/main/java/com/habit/server/controller/TeamController.java
@@ -1,0 +1,244 @@
+package com.habit.server.controller;
+
+import com.sun.net.httpserver.HttpExchange;
+import com.sun.net.httpserver.HttpHandler;
+import com.habit.server.AuthService;
+import com.habit.server.TeamRepository;
+import com.habit.server.UserRepository;
+import com.habit.domain.Team;
+import com.habit.domain.TeamMode;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
+
+/**
+ * チーム関連APIのコントローラ
+ */
+public class TeamController {
+    private final AuthService authService;
+    private final UserRepository userRepository;
+
+    public TeamController(AuthService authService, UserRepository userRepository) {
+        this.authService = authService;
+        this.userRepository = userRepository;
+    }
+
+    public HttpHandler getCreateTeamHandler() {
+        return new CreateTeamHandler();
+    }
+
+    public HttpHandler getJoinTeamHandler() {
+        return new JoinTeamHandler();
+    }
+
+    public HttpHandler getPublicTeamsHandler() {
+        return new PublicTeamsHandler();
+    }
+
+    public HttpHandler getFindTeamByPasscodeHandler() {
+        return new FindTeamByPasscodeHandler();
+    }
+
+    public HttpHandler getGetTeamNameHandler() {
+        return new GetTeamNameHandler();
+    }
+
+    // --- チーム作成API ---
+    class CreateTeamHandler implements HttpHandler {
+        @Override
+        public void handle(HttpExchange exchange) throws IOException {
+            if (!"POST".equals(exchange.getRequestMethod())) {
+                String response = "POSTメソッドのみ対応";
+                exchange.sendResponseHeaders(405, response.getBytes().length);
+                OutputStream os = exchange.getResponseBody();
+                os.write(response.getBytes());
+                os.close();
+                return;
+            }
+            byte[] bodyBytes = exchange.getRequestBody().readAllBytes();
+            String bodyStr = (bodyBytes != null) ? new String(bodyBytes, java.nio.charset.StandardCharsets.UTF_8) : "";
+            String response;
+            if (bodyStr == null || bodyStr.trim().isEmpty()) {
+                response = "リクエストボディが空です";
+                exchange.sendResponseHeaders(400, response.getBytes().length);
+                OutputStream os = exchange.getResponseBody();
+                os.write(response.getBytes());
+                os.close();
+                return;
+            }
+            try {
+                String[] params = bodyStr.split("&");
+                String teamID = UUID.randomUUID().toString();
+                String teamName = "", passcode = "", editPerm = "", category = "", scope = "public";
+                int maxMembers = 5;
+                List<String> members = new ArrayList<>();
+                for (String param : params) {
+                    String[] kv = param.split("=", 2);
+                    if (kv.length < 2) continue;
+                    switch (kv[0]) {
+                        case "teamName": teamName = java.net.URLDecoder.decode(kv[1], "UTF-8"); break;
+                        case "passcode": passcode = java.net.URLDecoder.decode(kv[1], "UTF-8"); break;
+                        case "maxMembers": maxMembers = Integer.parseInt(kv[1]); break;
+                        case "editPermission": editPerm = java.net.URLDecoder.decode(kv[1], "UTF-8"); break;
+                        case "category": category = java.net.URLDecoder.decode(kv[1], "UTF-8"); break;
+                        case "scope": scope = java.net.URLDecoder.decode(kv[1], "UTF-8"); break;
+                        case "members": for (String m : kv[1].split(",")) if (!m.isEmpty()) members.add(m); break;
+                    }
+                }
+                Team team = new Team(teamID, teamName, "creator", TeamMode.FIXED_TASK_MODE);
+                team.setteamName(teamName);
+                TeamRepository repo = new TeamRepository();
+                repo.save(team, passcode, maxMembers, editPerm, category, scope, members);
+
+                String sessionId = null;
+                var headers = exchange.getRequestHeaders();
+                if (headers.containsKey("SESSION_ID")) {
+                    sessionId = headers.getFirst("SESSION_ID");
+                }
+                if (sessionId == null) {
+                    String query = exchange.getRequestURI().getQuery();
+                    if (query != null && query.contains("SESSION_ID=")) {
+                        for (String param : query.split("&")) {
+                            if (param.startsWith("SESSION_ID=")) {
+                                sessionId = param.substring("SESSION_ID=".length());
+                                break;
+                            }
+                        }
+                    }
+                }
+                if (sessionId != null) {
+                    var user = authService.getUserBySession(sessionId);
+                    if (user != null) {
+                        user.addJoinedTeamId(teamID);
+                        userRepository.save(user);
+                    }
+                }
+                response = "チーム作成成功";
+            } catch (Exception ex) {
+                response = "チーム作成失敗: " + ex.getMessage();
+            }
+            exchange.sendResponseHeaders(200, response.getBytes().length);
+            OutputStream os = exchange.getResponseBody();
+            os.write(response.getBytes());
+            os.close();
+        }
+    }
+
+    // --- チーム参加API ---
+    class JoinTeamHandler implements HttpHandler {
+        @Override
+        public void handle(HttpExchange exchange) throws IOException {
+            String response;
+            String query = exchange.getRequestURI().getQuery();
+            String teamName = null;
+            if (query != null) {
+                String[] params = query.split("&");
+                for (String param : params) {
+                    if (param.startsWith("teamName=")) teamName = java.net.URLDecoder.decode(param.substring(9), "UTF-8");
+                }
+            }
+            String sessionId = null;
+            var headers = exchange.getRequestHeaders();
+            if (headers.containsKey("SESSION_ID")) {
+                sessionId = headers.getFirst("SESSION_ID");
+            }
+            if (teamName == null || teamName.isEmpty() || sessionId == null) {
+                response = "パラメータまたはセッションIDが不正です";
+            } else {
+                var user = authService.getUserBySession(sessionId);
+                if (user == null) {
+                    response = "ユーザが見つかりません";
+                } else {
+                    String memberId = user.getUserId();
+                    TeamRepository repo = new TeamRepository();
+                    boolean ok = repo.addMemberByTeamName(teamName, memberId);
+                    if (ok) {
+                        String teamID = repo.findTeamIdByName(teamName);
+                        if (teamID != null) {
+                            user.addJoinedTeamId(teamID);
+                            userRepository.save(user);
+                        }
+                    }
+                    response = ok ? "参加成功" : "参加失敗";
+                }
+            }
+            exchange.sendResponseHeaders(200, response.getBytes().length);
+            OutputStream os = exchange.getResponseBody();
+            os.write(response.getBytes());
+            os.close();
+        }
+    }
+
+    // --- 公開チーム一覧取得API ---
+    class PublicTeamsHandler implements HttpHandler {
+        @Override
+        public void handle(HttpExchange exchange) throws IOException {
+            String response;
+            try {
+                TeamRepository repo = new TeamRepository();
+                List<String> teamNames = repo.findAllPublicTeamNames();
+                response = String.join("\n", teamNames);
+            } catch (Exception ex) {
+                response = "エラー: " + ex.getMessage();
+            }
+            exchange.sendResponseHeaders(200, response.getBytes().length);
+            OutputStream os = exchange.getResponseBody();
+            os.write(response.getBytes());
+            os.close();
+        }
+    }
+
+    // --- 合言葉でチーム検索API ---
+    class FindTeamByPasscodeHandler implements HttpHandler {
+        @Override
+        public void handle(HttpExchange exchange) throws IOException {
+            String response;
+            String query = exchange.getRequestURI().getQuery();
+            String passcode = null;
+            if (query != null && query.startsWith("passcode=")) {
+                passcode = java.net.URLDecoder.decode(query.substring(9), "UTF-8");
+            }
+            if (passcode == null || passcode.isEmpty()) {
+                response = "合言葉が指定されていません";
+            } else {
+                TeamRepository repo = new TeamRepository();
+                String teamName = repo.findTeamNameByPasscode(passcode);
+                if (teamName != null) {
+                    response = teamName;
+                } else {
+                    response = "該当チームなし";
+                }
+            }
+            exchange.sendResponseHeaders(200, response.getBytes().length);
+            OutputStream os = exchange.getResponseBody();
+            os.write(response.getBytes());
+            os.close();
+        }
+    }
+    // --- チーム名取得API ---
+    class GetTeamNameHandler implements HttpHandler {
+        @Override
+        public void handle(HttpExchange exchange) throws IOException {
+            String query = exchange.getRequestURI().getQuery();
+            String teamID = null;
+            if (query != null && query.startsWith("teamID=")) {
+                teamID = java.net.URLDecoder.decode(query.substring(7), "UTF-8");
+            }
+            String response;
+            if (teamID == null || teamID.isEmpty()) {
+                response = "";
+            } else {
+                TeamRepository repo = new TeamRepository();
+                String name = repo.findTeamNameById(teamID);
+                response = (name != null) ? name : "";
+            }
+            exchange.getResponseHeaders().set("Content-Type", "text/plain; charset=UTF-8");
+            exchange.sendResponseHeaders(200, response.getBytes("UTF-8").length);
+            OutputStream os = exchange.getResponseBody();
+            os.write(response.getBytes("UTF-8"));
+            os.close();
+        }
+    }
+}
